Amended any (Integer) criteriaUniqueResult() to Long(s) (avoids class cas exceptions if/when Hibernate is upgraded to later version)
git-svn-id: svn://localhost/ARK@6465 35ee9b83-dd2c-47f4-99a2-c706670c48ce

+++ ark-common/src/main/java/au/org/theark/core/dao/ArkAuthorisationDao.java
@@ -1064,7 +1064,7 @@ public class ArkAuthorisationDao<T> extends HibernateSessionDao implements IArkA
 		Criteria criteria = getSession().createCriteria(ArkUserRole.class);
 		criteria.add(Restrictions.eq("arkUser", arkUser));
 		criteria.setProjection(Projections.rowCount());
-		Integer totalCount = (Integer) criteria.uniqueResult();
+		Long totalCount = (Long) criteria.uniqueResult();
 		if (totalCount > 0) {
 			flag = true;
 		}
